+++ packages/pwa/index.html
@@ -9,7 +9,7 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/index.ts"></script>
+    <script type="module" src="/src/index.tsx"></script>
     <script type="text/typescript">
       if (navigator.serviceWorker) {
         navigator.serviceWorker
