+++ packages/shared/src/jsx/jsx-dev-runtime.ts
@@ -1,3 +1,5 @@
-import { jsx } from './jsx-runtime';
+import { Fragment as F, jsx } from './jsx-runtime';
 
 export const jsxDEV = jsx;
+export const jsxsDEV = jsx;
+export const Fragment = F;
